Add Interfaces for Lexicon, DocumentIndex and reduce memory size of program
- Decrease memory size of Lexicon
- Change search method for use WordID (not Word Text)
- Remove Word Text from memory and from lexicon
- Add web interface for test
- Fix: First version of ranked query (doesn't consider phrase search or
proximity or words (next version?)

+++ DocCore/Word/WordOccurrenceNode.cs
@@ -41,8 +41,7 @@ namespace DocCore
         {
             get {
                 
-                return hits.Count / doc.WordQuantity; 
-                
+                return ((double) hits.Count) / ((double)doc.WordQuantity); 
             }
         }
